Refactored backpressure relief.
Pulls out downlink specific code.
